+++ app/src/main/java/com/codewithre/simedit/data/remote/response/BalanceResponse.kt
@@ -1,6 +1,7 @@
 package com.codewithre.simedit.data.remote.response
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigInteger
 
 data class BalanceResponse(
 
@@ -8,7 +9,7 @@ data class BalanceResponse(
 	val code: Int? = null,
 
 	@field:SerializedName("data")
-	val data: Int? = null,
+	val data: BigInteger? = null,
 
 	@field:SerializedName("success")
 	val success: Boolean? = null,
